migrate rune
