+++ languages/bevy_mod_scripting_rune/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "bevy_mod_scripting_rune"
-version = "0.3.0"
+version = "0.4.0"
 edition = "2021"
 license = "MIT"
 description = "Necessary functionality for Rune support with bevy_mod_scripting"
@@ -21,7 +21,8 @@ name = "bevy_mod_scripting_rune"
 path = "src/lib.rs"
 
 [dependencies]
-bevy_mod_scripting_core = { path = "../../bevy_mod_scripting_core", version = "0.3.0" }
-bevy = "0.11"
+bevy_mod_scripting_core = { path = "../../bevy_mod_scripting_core", version = "0.4.0" }
+bevy = "0.12.1"
 rune = "0.13.1"
 rune-modules = "0.13.1"
+anyhow = "1.0.75"
